Use flake8 for Python linting instead of pyflakes and pycodestyle
flake8 is a wrapper around pyflakes, pycodestyle, and mccabe, and also
can be further extended using third-party plugins.  The Travis-CI
tests were already running pyflakes and pycodestyle.  We may as well
unify both of those under the flake8 umbrella so that we have a single
place to manage configuration, unified output of any complaints, etc.

By the way, bravo to the canvasapi developers for maintaining such a
clean code base!