Add checks for `context_code` key to announcements
Announcements come in through the `Canvas` object and do not include
a course ID directly. Since announcements are DiscussionTopics, we
can parse the ID from the `context_code` param.

Currently, the Canvas API only supports course-level context codes,
so this works for now, but will break as more contexts are added.

https://canvas.instructure.com/doc/api/announcements.html#method.announcements_api.index