Accept a list of course IDs in get_announcements
Canvas returns announcements from the `canvas` object. The user has to
define which course or courses to pull from. Adding `course_list` as a
required parameter prevents failed calls by exposing this required
parameter to the user.

The list is processed to the `context_codes` format before being sent
in the request.