Better custom gradebook column CRUD
Without the `{"content": column_data}`, we get a nasty `500` error everytime this algorithm tries to execute. I was surprised to hit it the first time I executed my script. The other change is to be able to set the column data for the first time. When there is already some data in the custom column, the `get_entries` routine returns and then we can modify then. But if there's nothing there in the first place, it is impossible with the current code base to add anything to start with. Therefore I propose to add in the `CustomGradebookColumn` a update the column data; this way it will be possible to create entries as well as modify them.