change update_score_and_comments to wrap args
when calling update_score_and_comments on a quiz submission, the
attempt is already in self and there will be only one item in the
update list.

this change wraps the passed arguments in a unary quiz_submissions
list if quiz_submissions is not one of the parameters to the call.

legacy behavior is preserved if quiz_submissions is in the calling
parameters.