Use `new_quizzes` and `graphql` keywords
Based on #612 and as an issue in #619, this adds the `new_quizzes` and
`graphql` keyword endpoints in the requester module. Methods which
specify these locations will have the correct URL supplied by the
library in the `_url` parameter for the requester.

Calls from the `canvas`, `course`, and `new_quizzes` modules have been
updated to use they keywords rather than a formatted URL string.

All tests passing.