Param for get_appointment_groups, checks for creating new appointment groups
Fixes #678

Two issues in one:

1. `canvas.get_appointment_groups()` returns an empty list for teachers
because it defaults to "reservable", but instructors can't reserve, so
the list is empty. Create a new param on the method that defaults to
`reservable` but can be overridden. Update the docstring to note the
change.

2. Update `create_appointment_group()` to take either a course ID or a
`Course` object. Check that the `context_codes` key is present on the
`appointment_group` dict before checking formatting.

All tests passing.