fix: don't override `per_page` in `PaginatedList`
Currently a default is set for `per_page` to 100, but later the manner of passing params to the functions was updated such that it is not the default but instead overrides `per_page`.
You would get a dict of `{"_kwargs": [("per_page": 1)], "per_page": 100}
which then converts to the following request
https://institution/api/v1/courses?per_page=1&per_page=100
since we define `per_page` later it will also be put in later when expanded with .items().

keeps the same behaviour as before except that _kwargs now must be list[tuple[str,Any]], this doesn't break anything internally, but can break other people's code.

Also made _kwargs keyword consistent for all PaginatedList calls, couldn't find any reason why they were different.